tacit-knowledge/worker/pyproject.toml
[project]
name = "skillrelay-python-server"
version = "1.0.0"
description = "SkillRelay Python Server - Document Search API & Worker"
requires-python = ">=3.9"
dependencies = [
    "llama-index-core>=0.14.10",
    "llama-index-vector-stores-postgres",
    "llama-index-embeddings-openai",
    "llama-index-llms-openai",
    "llama-index-readers-file",
    "llama-index-readers-whisper",
    "llama-index-storage-chat-store-postgres",
    "s3fs",
    "python-dotenv",
    "boto3",
    "pypdf",
    "PyMuPDF",
    "pdf2image",
    "psycopg2-binary",
    "pyyaml",
    "openai",
    "numpy",
    "pydantic",
    "tiktoken",
    "docx2txt",
    "python-pptx",
    "openpyxl",
    "python-ulid",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
